requirements.txt
django == 1.8.7 
django-mmc == 1.6
django-admin-bootstrapped == 2.5.6
django-app-namespace-template-loader == 0.3
django-bootstrap3-datetimepicker-2 == 2.2.4.2

django-extensions == 1.5.9
pygraphviz == 1.3.1
django-debug-toolbar == 1.4
